# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  apierror.ErrorResponse:
    properties:
      error:
        description: Stable error code (e.g., "invalid_credentials")
        type: string
      message:
        description: Human-readable message
        type: string
    type: object
...
  dto.CreateTeamTypeRequest:
    properties:
      code:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
...
  handler.agentRolePromptResponse:
    properties:
      content:
        type: string
      description:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  handler.mcpServerRegistryRequest:
    type: object
...
  handler.updateRolePromptRequest:
    properties:
      content:
        type: string
      description:
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Backend API с авторизацией на JWT токенах
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
  /admin/agent-role-prompts:
    get:
      description: Возвращает реестр дефолтных системных промптов для каждой роли
        агента. Используется в админке для просмотра и редактирования.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.agentRolePromptResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List default role prompts
      tags:
      - admin-role-prompts
  /admin/agent-role-prompts/{role}:
    get:
      description: Возвращает дефолтный системный промпт для указанной роли агента.
      parameters:
      - description: Роль агента (assistant, orchestrator, router, planner, ...)
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.agentRolePromptResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get role prompt by role
      tags:
      - admin-role-prompts
    put:
      consumes:
      - application/json
      description: Обновляет content и description дефолтного промпта. Только для
        admin. Сброс к дефолту = удалить запись в БД + перезапустить backend (seed
        пересоздаст).
      parameters:
      - description: Роль агента
        in: path
        name: role
        required: true
        type: string
      - description: Новый промпт
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateRolePromptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.agentRolePromptResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update role prompt
      tags:
      - admin-role-prompts
  /admin/mcp-servers:
    get:
      parameters:
      - description: Only active servers
        in: query
        name: only_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MCPServerRegistry'
            type: array
      security:
      - BearerAuth: []
      summary: List MCP servers
      tags:
      - admin-mcp-servers
    post:
      consumes:
      - application/json
      parameters:
      - description: Server definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.mcpServerRegistryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MCPServerRegistry'
      security:
      - BearerAuth: []
      summary: Create MCP server
      tags:
      - admin-mcp-servers
  /admin/mcp-servers/{id}:
    delete:
      parameters:
      - description: Server UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete MCP server
      tags:
      - admin-mcp-servers
    get:
      parameters:
      - description: Server UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MCPServerRegistry'
      security:
      - BearerAuth: []
      summary: Get MCP server
      tags:
      - admin-mcp-servers
    put:
      consumes:
      - application/json
      parameters:
      - description: Server UUID
        in: path
        name: id
        required: true
        type: string
      - description: Server definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.mcpServerRegistryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MCPServerRegistry'
      security:
      - BearerAuth: []
      summary: Update MCP server
      tags:
      - admin-mcp-servers
  /admin/team-types:
    post:
      consumes:
      - application/json
      description: Создает новый тип команды (только для администраторов)
      parameters:
      - description: Параметры создания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTeamTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TeamTypeResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      - OAuth2Password: []
      summary: Создание типа команды
      tags:
      - teams
  /admin/team-types/{code}:
    delete:
      description: Удаляет тип команды по коду (только для администраторов)
      parameters:
      - description: Код типа команды
        in: path
        name: code
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      - OAuth2Password: []
      summary: Удаление типа команды
      tags:
      - teams
  /agents:
    get:
      description: Реестр LLM/sandbox-агентов с фильтрами (only_active, execution_kind,
        role, name_like) и пагинацией. system_prompt НЕ включается; полную запись
        — через GET /agents/:id.
      parameters:
      - description: только is_active=true
        in: query
        name: only_active
        type: boolean
      - description: llm | sandbox
        in: query
        name: execution_kind
        type: string
      - description: router/planner/decomposer/reviewer/developer/merger/tester/...
        in: query
        name: role
        type: string
      - description: частичный поиск по name
        in: query
        name: name_like
        type: string
      - description: 1-200; default 50
        in: query
        name: limit
        type: integer
      - description: default 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List v2 agents
      tags:
      - agents-v2
    post:
      consumes:
      - application/json
      parameters:
      - description: agent payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createAgentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Agent'
        "400":
          description: validation
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "409":
          description: name already taken
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create v2 agent
      tags:
      - agents-v2
...
  /agents/{id}/secrets:
    post:
      consumes:
      - application/json
      description: Шифрует value через AES-256-GCM (AAD = secret.id), сохраняет в
        agent_secrets. Back-read невозможен через API.
      parameters:
      - description: Agent UUID
        in: path
        name: id
        required: true
        type: string
      - description: key_name + plaintext value
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.setSecretRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.SetSecretOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set v2 agent secret (encrypted)
      tags:
      - agents-v2
...
  /agents/secrets/{secret_id}:
    delete:
      parameters:
      - description: Secret UUID
        in: path
        name: secret_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete v2 agent secret
      tags:
      - agents-v2
  /antigravity/auth:
    delete:
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Отзыв OAuth-подписки Antigravity
      tags:
      - antigravity
...
  /auth/me:
    get:
      consumes:
      - application/json
      description: Возвращает данные аутентифицированного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      - OAuth2Password: []
      summary: Получение данных текущего пользователя
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Long-lived API key for programmatic access. Format: wibe_<key>'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  OAuth2Password:
    flow: password
    tokenUrl: /api/v1/auth/login
    type: oauth2
swagger: "2.0"